requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras requirements.in
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   starlette
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
fastapi==0.110.0
    # via -r requirements.in
google-auth==2.28.2
    # via google-auth-oauthlib
google-auth-oauthlib==1.2.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via -r requirements.in
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.3
    # via -r requirements.in
mako==1.3.2
    # via alembic
markupsafe==2.1.5
    # via
    #   jinja2
    #   mako
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.2
    # via gunicorn
passlib==1.7.4
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.6.3
    # via
    #   fastapi
    #   pydantic-settings
pydantic-core==2.16.3
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via pydantic-settings
python-multipart==0.0.9
    # via -r requirements.in
requests==2.31.0
    # via requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.28
    # via
    #   -r requirements.in
    #   alembic
starlette==0.36.3
    # via fastapi
typing-extensions==4.10.0
    # via
    #   alembic
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.2.1
    # via requests
uvicorn==0.28.0
    # via -r requirements.in
